change from long to int

+++ a.java
@@ -5,8 +5,8 @@ import java.io.*;
 public class a {
     public static void main(String arg[]){
         Scanner sc = new Scanner(System.in);
-        long n= sc.nextLong();
-        long factorial= 1;
+        int  n= sc.nextInt();
+        int factorial= 1;
         for(int i=1;i<=n;i++)
         {
             factorial=factorial*i;
